// src/renderer/shared/models/enums/ConditionInitiator.enum.ts
export enum ConditionInitiator {
    UNDEFINED = 'model.undefined',

    ENTITY_FILTERING = 'model.condition.intiator.entity_filter',
    TRAIT = 'model.condition.initiator.trait',
    EVENT_FLAGGED = 'model.condition.initiator.event_flag',
    LOCATION = 'model.condition.initiator.location',
    TIME = 'model.condition.initiator.time',
    RELATIONSHIP = 'model.condition.initiator.relationship',
}